--- src/components/common/loading/loadingPage.module.scss
@@ -1,46 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-    background-color: #fff;
-    padding: 40px 0;
-}
-  
-.content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    flex-grow: 1;
-    justify-content: center;
-}
-  
-.logo {
-    margin-bottom: 24px;
-}
-
-.title {
-    align-self: stretch;
-    color: var(--grey-grey-09, #424242);
-    text-align: center;
-    font-family: Pretendard;
-    font-size: 32px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: normal;
-    margin-bottom: 8px;
-}
-
-.subtitle {
-    align-self: stretch;
-    color: var(--grey-grey-07, #757575);
-    text-align: center;
-    font-family: Pretendard;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-    margin-bottom: 32px;
-}
