+++ src/components/common/modal/RequestSupportModal.tsx
@@ -1,17 +1,18 @@
 import { useEffect } from 'react';
 import styles from './supportModal.module.scss';
-import ArrowIcon from '../../assets/icons/arrowUp2Thin.svg?react';
-import CheckLightIcon from '../../assets/icons/checkLight.svg?react';
-import ApplyIcon from '../../assets/icons/apply.svg?react';
-import UnapplyIcon from '../../assets/icons/unapply.svg?react';
-import { getCategoryLabel } from '../../utils/categoryMapping';
-import { useMutation } from '@tanstack/react-query';
-import { applyForCollaboration } from '../../apis/applyAPI';
 import axios from 'axios';
-import { useCollaborationSupport } from '../../hooks/useCollaborationSupport';
-import { useToast } from '../../contexts/toastContext';
+import { useMutation } from '@tanstack/react-query';
 
-interface CollaborationSupportModalProps {
+import UnapplyIcon from '../../../assets/icons/unapply.svg?react';
+import ApplyIcon from '../../../assets/icons/apply.svg?react';
+import CheckLightIcon from '../../../assets/icons/checkLight.svg?react';
+import ArrowIcon from '../../../assets/icons/arrowUp2Thin.svg?react';
+import { getCategoryLabel } from '../../../utils/categoryMapping';
+import { applyForRequest } from '../../../apis/applyAPI';
+import { useCollaborationSupport } from '../../../hooks/useCollaborationSupport';
+import { useToast } from '../../../contexts/toastContext';
+
+interface RequestSupportModalProps {
   onClose: () => void;
   recruitments: {
     recruitmentId: number;
@@ -22,17 +23,17 @@ interface CollaborationSupportModalProps {
   category: string;
   writerName: string;
   title: string;
-  collaborationId: number;
+  taskId: number;
 }
 
-export const CollaborationSupportModal = ({
+const RequestSupportModal = ({
   onClose,
   recruitments,
   category,
   writerName,
   title,
-  collaborationId,
-}: CollaborationSupportModalProps) => {
+  taskId,
+}: RequestSupportModalProps) => {
   const {
     selectedSupport,
     message,
@@ -53,9 +54,9 @@ export const CollaborationSupportModal = ({
 
   const mutation = useMutation({
     mutationFn: async () => {
-      if (!collaborationId) {
-        errorToast('협업 게시글 ID가 없습니다. 다시 시도해주세요.');
-        return Promise.reject('collaborationId is missing');
+      if (!taskId) {
+        errorToast('요청과제 게시글 ID가 없습니다. 다시 시도해주세요.');
+        return Promise.reject('taskId is missing');
       }
 
       if (!selectedSupport) {
@@ -64,17 +65,16 @@ export const CollaborationSupportModal = ({
       }
 
       const requestData = {
-        collaborationRecruitmentId: selectedSupport,
+        requestRecruitmentId: selectedSupport,
         isOpenProfile: isChecked,
         message,
       };
-
-      return applyForCollaboration(collaborationId, requestData);
+      return applyForRequest(taskId, requestData);
     },
     onSuccess: () => {
       successToast('지원이 완료되었습니다! 지원 결과를 기다려주세요');
       onClose();
-      navigate('/my/apply-collaboration');
+      navigate('/my/apply-request');
     },
     onError: (error: unknown) => {
       let errorMessage = '지원 요청 중 오류가 발생했습니다.';
@@ -99,13 +99,13 @@ export const CollaborationSupportModal = ({
         className={styles.modal}
         onClick={(e) => e.stopPropagation()}>
         <div className={styles.header}>
-          <h1 className={styles.title}>협업 광장 지원</h1>
+          <h1 className={styles.title}>요청과제 지원</h1>
         </div>
         <hr className={styles.divider} />
         <div className={styles.content}>
           <div className={styles.titleSection}>
             <div className={styles.breadcrumb}>
-              <span>협업 광장</span>
+              <span>요청과제</span>
               <ArrowIcon className={styles.arrowIcon} />
               <span>{getCategoryLabel(category)}</span>
             </div>
@@ -115,7 +115,6 @@ export const CollaborationSupportModal = ({
             </div>
           </div>
           <h2 className={styles.sectionTitle}>지원 부문</h2>
-          <hr className={styles.divider} />
           {recruitments.length === 0 ? (
             <p className={styles.noRecruitments}>
               지원할 모집 부문이 없습니다.
@@ -166,9 +165,7 @@ export const CollaborationSupportModal = ({
                 <CheckLightIcon className={styles.checkIcon} />
               </div>
               <div
-                className={`${styles.checkboxLabel} ${
-                  isChecked ? styles.checkedLabel : ''
-                }`}>
+                className={`${styles.checkboxLabel} ${isChecked ? styles.checkedLabel : ''}`}>
                 프로필 공개
               </div>
             </div>
@@ -198,3 +195,5 @@ export const CollaborationSupportModal = ({
     </div>
   );
 };
+
+export default RequestSupportModal;
