+++ src/components/error/LoadingSpinner.tsx
@@ -4,7 +4,7 @@ import styles from './loadingSpinner.module.scss';
 const LoadingSpinner = () => {
   return (
     <div className={styles.spinner}>
-      <PulseLoader />
+      <PulseLoader color='#377FF8' />
     </div>
   );
 };
