+++ src/components/error/loadingSpinner.module.scss
@@ -3,9 +3,10 @@
   align-items: center;
   gap: 15px;
 
+  /* stylelint-disable-next-line selector-class-pattern */
   .react-spinners-PulseLoader {
     span {
-      background-color: $primary
+      background-color: $primary;
     }
   }
 }
